Refactored LoHaLayer and LoKrLayer a bit
