+++ src/peft/tuners/lokr/layer.py
@@ -14,7 +14,6 @@
 # limitations under the License.
 
 import math
-import warnings
 from itertools import chain
 from typing import Iterable, Optional, Tuple, Union
 
@@ -42,9 +41,6 @@ class LoKrLayer(LyCORISLayer, nn.Module):
         super(nn.Module, self).__init__()
 
         # LoKr info
-        self.r = {}
-        self.alpha = {}
-        self.scaling = {}
         self.lokr_w1 = nn.ParameterDict({})
         self.lokr_w1_a = nn.ParameterDict({})
         self.lokr_w1_b = nn.ParameterDict({})
@@ -52,20 +48,6 @@ class LoKrLayer(LyCORISLayer, nn.Module):
         self.lokr_w2_a = nn.ParameterDict({})
         self.lokr_w2_b = nn.ParameterDict({})
         self.lokr_t2 = nn.ParameterDict({})
-        self.rank_dropout = {}
-        self.module_dropout = {}
-
-    def _init_empty_weights(self, cls, *args, **kwargs) -> None:
-        # A helper method that allows to initialize the layer of the given class without spending time to initialize the
-        # model weights. The implementation is inspired by
-        # https://pytorch.org/docs/stable/generated/torch.nn.utils.skip_init.html but this function cannot be used
-        # directly.
-        # Instead of this approach, it would be possible to bypass the __init__ of the class but that runs the risk of
-        # omitting important logic inside that __init__.
-        kwargs = kwargs.copy()
-        final_device = kwargs.pop("device", "cpu")
-        cls.__init__(self, *args, device="meta", **kwargs)
-        self.to_empty(device=final_device)
 
     @property
     def _available_adapters(self) -> Iterable[str]:
@@ -81,7 +63,7 @@ class LoKrLayer(LyCORISLayer, nn.Module):
             )
         )
 
-    def create_lokr_parameters(
+    def create_adapter_parameters(
         self,
         adapter_name: str,
         r: int,
@@ -115,7 +97,7 @@ class LoKrLayer(LyCORISLayer, nn.Module):
                 self.lokr_w2_a[adapter_name] = nn.Parameter(torch.empty(shape[0][1], r))
                 self.lokr_w2_b[adapter_name] = nn.Parameter(torch.empty(r, shape[1][1]))
 
-    def reset_lokr_parameters(self, adapter_name: str):
+    def reset_adapter_parameters(self, adapter_name: str):
         if adapter_name in self.lokr_w1:
             nn.init.zeros_(self.lokr_w1[adapter_name])
         else:
@@ -188,11 +170,11 @@ class LoKrLayer(LyCORISLayer, nn.Module):
             raise NotImplementedError(f"LoKr is not implemented for {type(self).__name__} layer")
 
         # Create weights with provided shape
-        self.create_lokr_parameters(adapter_name, r, shape, use_w1, use_w2, use_effective_conv2d)
+        self.create_adapter_parameters(adapter_name, r, shape, use_w1, use_w2, use_effective_conv2d)
 
         # Initialize weights
         if init_weights:
-            self.reset_lokr_parameters(adapter_name)
+            self.reset_adapter_parameters(adapter_name)
 
         # Move new weights to device
         weight = getattr(self, "weight", None)
@@ -231,72 +213,6 @@ class LoKrLayer(LyCORISLayer, nn.Module):
             weight *= drop
 
         return weight
-
-    def merge(self) -> None:
-        if self.merged:
-            warnings.warn(
-                f"Already following adapters were merged {','.join(self.merged_adapters)}. "
-                f"You are now additionally merging {','.join(self.active_adapters)}."
-            )
-        for active_adapter in self.active_adapters:
-            if active_adapter in self._available_adapters:
-                self.weight.data += self.get_delta_weight(active_adapter)
-                self.merged_adapters.append(active_adapter)
-
-    def unmerge(self) -> None:
-        if not self.merged:
-            warnings.warn("Already unmerged. Nothing to do.")
-            return
-        while len(self.merged_adapters) > 0:
-            active_adapter = self.merged_adapters.pop()
-            if active_adapter in self._available_adapters:
-                self.weight.data -= self.get_delta_weight(active_adapter)
-
-    def _op(self, x: torch.Tensor, weight: torch.Tensor) -> torch.Tensor:
-        raise NotImplementedError
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        previous_dtype = x.dtype
-
-        if self.disable_adapters:
-            if self.merged:
-                self.unmerge()
-            result = self._op(x, self.weight)
-        elif self.merged:
-            result = self._op(x, self.weight)
-        else:
-            # Get base weights
-            weight = self.weight.data
-
-            # Execute all the adapters
-            for active_adapter in self.active_adapters:
-                if active_adapter not in self._available_adapters:
-                    continue
-
-                module_dropout = self.module_dropout[active_adapter]
-
-                # Modify current execution weights
-                if (not self.training) or (self.training and torch.rand(1) > module_dropout):
-                    weight = weight + self.get_delta_weight(active_adapter)
-
-            # Perform actual operation
-            result = self._op(x, weight)
-
-        result = result.to(previous_dtype)
-        return result
-
-    def scale_layer(self, scale_factor: float) -> None:
-        if scale_factor != 1:
-            for active_adapter in self.active_adapters:
-                alpha = self.alpha[active_adapter]
-                r = self.r[active_adapter]
-                self.scaling[active_adapter] = (alpha / r) * scale_factor
-
-    def unscale_layer(self) -> None:
-        for active_adapter in self.active_adapters:
-            alpha = self.alpha[active_adapter]
-            r = self.r[active_adapter]
-            self.scaling[active_adapter] = alpha / r
 
 
 class Linear(LoKrLayer, nn.Linear):
